patch 9.2.1077: GUI: gui_mch_wait_for_chars() takes an int on some systems
Problem:  gui_wait_for_chars() passes a long to gui_mch_wait_for_chars(),
          the GTK and X11 implementations take a long, the MS-Windows,
          Photon and Haiku ones take an int.
Solution: Take a long everywhere (Hirohito Higashi).

closes: #21278

+++ src/gui_w32.c
@@ -2666,7 +2666,7 @@ server_flush_input(void)
  * or FAIL otherwise.
  */
     int
-gui_mch_wait_for_chars(int wtime)
+gui_mch_wait_for_chars(long wtime)
 {
     int		focus;
